chore: convert LESS to SCSS, move SCSS next to component files, organize better infinite slider code [PT-188049278]
