+++ package.json
@@ -41,7 +41,6 @@
     "autoprefixer": "^10.4.19",
     "css-loader": "^7.1.2",
     "html-webpack-plugin": "^5.6.0",
-    "less-loader": "^12.2.0",
     "postcss-loader": "^8.1.1",
     "sass": "^1.77.8",
     "sass-loader": "^16.0.0",
