+++ src/components/view-manager.scss
