+++ src/components/view-manager.tsx
@@ -5,7 +5,7 @@ import RaysViewComp from './rays-view-comp';
 import t from '../translation/translate';
 import { ISimState, IViewState, ViewType } from '../types';
 
-import '../css/view-manager.less';
+import './view-manager.scss';
 
 interface IProps {
   simulation: ISimState;
