--- src/css/jquery-ui-theme.less
@@ -1,30 +1,0 @@
-.grasp-seasons {
-  // Note that jquery-ui theme imports are scoped to .grasp-seasons, so we won't pollute rest of the page.
-  // Use (less) to enforce LESS processing. Otherwise, style won't respect scope nesting.
-  @import (less) "../../node_modules/jquery-ui/themes/base/slider.css";
-
-  .ui-widget {
-    font-family: inherit;
-    font-size: inherit;
-  }
-
-  .ui-slider{
-    background-color: #fff;
-    border: 1px solid #999;
-    border-radius:4px;
-  }
-
-  .ui-slider-handle {
-    border-radius: 50px;
-    background: #fff !important;
-    box-shadow: 0 0 6px #666;
-  }
-
-  .ui-slider-with-tick-labels {
-    margin-bottom: 25px;
-  }
-
-  .ui-slider-tick-label {
-    margin-top: 2px;
-  }
-}
