+++ src/global.d.ts
@@ -1,3 +1,4 @@
 declare module "*.jpg";
 declare module "*.png";
 declare module "*.svg";
+declare module "*.scss";
