+++ webpack.config.js
@@ -48,19 +48,6 @@ module.exports = {
         }]
       },
       {
-        test: /\.less$/,
-        use: [{
-          loader: 'style-loader' // creates style nodes from JS strings
-        }, {
-          loader: 'css-loader', // translates CSS into CommonJS
-          options: {
-            esModule: false
-          }
-        }, {
-          loader: 'less-loader' // compiles Less to CSS
-        }]
-      },
-      {
         test: /\.(scss)$/,
         use: [{
           loader: 'style-loader', // inject CSS to page
